feat: 게시글 상세 페이지 좋아요 기능 구현
홈 화면 게시글 좋아요 와 비슷한 로직으로 상세페이지에도 좋아요를 만들어
주었습니다. 좋아요의 개수는 실시간으로 반영되며 상세페이지와 홈 화면
좋아요는 서로 연결되어 있습니다. 계정 마다 1개의 게시글의 최대 1개 할 수 있으며 true false로 구분지어 놨습니다.
계정이 로그인 되지 않은 상태에서 좋아요 클릭 시 로그인 페이지로
이동합니다.
