+++ src/components/Like.jsx
@@ -1,19 +1,21 @@
 import React, { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { styled } from 'styled-components';
-import { db, auth } from '../firebase';
-import { collection, query, onSnapshot, orderBy, getDocs, updateDoc, doc, setDoc } from 'firebase/firestore';
-import { Link } from 'react-router-dom';
-import Layout from './Layout';
+import { db } from '../firebase';
+import { collection, query,getDocs, updateDoc, doc, setDoc } from 'firebase/firestore';
+
 import { useNavigate } from 'react-router-dom';
 import uuid from 'react-uuid';
-
-const Like = () => {
-  const [reload, setReload] = useState(false);
+const Like = ({ item }) => {
+  const navigate = useNavigate();
   const [likeStatus, setLikeStatus] = useState(false);
   const [like, setLike] = useState([]);
+  const { storeInfo } = useSelector(state => state.userLogIn);
+ 
 
+  // 데이터 리스트로 불러오기
   useEffect(() => {
+    console.log('Like');
     const likesAllData = [];
     const fetchData = async () => {
       const likesValue = query(collection(db, 'likes'));
@@ -26,33 +28,41 @@ const Like = () => {
         likesAllData.push(data);
       });
       const likesData = storeInfo.uid ? likesAllData.filter(e => storeInfo.uid === e.uid) : [];
+      setLike(likesData);
     };
-    setLike(likesData);
     fetchData();
-  }, [reload, likeStatus]);
+  }, [likeStatus]);
 
   const postLike = async (item, likeOrNot, likeid) => {
-    const UUID = uuid();
-    const postRef = doc(db, 'post-item', item.postId);
-    const collectionRef = doc(collection(db, 'likes'), UUID);
-    let searchRef;
-    likeid ? (searchRef = doc(collection(db, 'likes'), likeid)) : (searchRef = false);
-    if (!likeOrNot) {
-      await updateDoc(postRef, { ...item, likeCount: item.likeCount + 1 });
-      searchRef
-        ? updateDoc(searchRef, { likeOrNot: !likeOrNot })
-        : setDoc(collectionRef, {
-            id: UUID,
-            uid: auth.currentUser.uid,
-            postId: item.postId,
-            likeOrNot: true,
-          });
-    } else if (likeOrNot) {
-      await updateDoc(postRef, { ...item, likeCount: item.likeCount - 1 });
-      updateDoc(searchRef, { likeOrNot: !likeOrNot });
+    if (!storeInfo.uid) {
+      alert('좋아요 기능은 로그인을 해야 가능합니다!');
+      const confirm = window.confirm('로그인 하시겠습니까?');
+      if (confirm) navigate('/login');
+    } else {
+      const UUID = uuid();
+      const postRef = doc(db, 'post-item', item.postId);
+      const collectionRef = doc(collection(db, 'likes'), UUID);
+      let searchRef;
+      likeid ? (searchRef = doc(collection(db, 'likes'), likeid)) : (searchRef = false);
+      if (!likeOrNot) {
+        await updateDoc(postRef, { ...item, likeCount: item.likeCount + 1 });
+        searchRef
+          ? await updateDoc(searchRef, { likeOrNot: !likeOrNot })
+          : await setDoc(collectionRef, {
+              id: UUID,
+              uid: storeInfo.uid,
+              postId: item.postId,
+              likeOrNot: true,
+            });
+      } else if (likeOrNot) {
+        await updateDoc(postRef, { ...item, likeCount: item.likeCount - 1 });
+        await updateDoc(searchRef, { likeOrNot: !likeOrNot });
+      }
+      setLikeStatus(!likeStatus);
     }
-    setLikeStatus(!likeStatus);
   };
+  const likeOrNot = like.find(e => e.postId === item.postId)?.likeOrNot ?? false;
+  const likeid = like.find(e => e.postId === item.postId)?.id ?? false;
   return (
     <HeartDiv onClick={() => postLike(item, likeOrNot, likeid)} style={{ cursor: 'pointer' }}>
       {likeOrNot ? (
@@ -64,15 +74,21 @@ const Like = () => {
           <path d="M225.8 468.2l-2.5-2.3L48.1 303.2C17.4 274.7 0 234.7 0 192.8v-3.3c0-70.4 50-130.8 119.2-144C158.6 37.9 198.9 47 231 69.6c9 6.4 17.4 13.8 25 22.3c4.2-4.8 8.7-9.2 13.5-13.3c3.7-3.2 7.5-6.2 11.5-9c0 0 0 0 0 0C313.1 47 353.4 37.9 392.8 45.4C462 58.6 512 119.1 512 189.5v3.3c0 41.9-17.4 81.9-48.1 110.4L288.7 465.9l-2.5 2.3c-8.2 7.6-19 11.9-30.2 11.9s-22-4.2-30.2-11.9zM239.1 145c-.4-.3-.7-.7-1-1.1l-17.8-20c0 0-.1-.1-.1-.1c0 0 0 0 0 0c-23.1-25.9-58-37.7-92-31.2C81.6 101.5 48 142.1 48 189.5v3.3c0 28.5 11.9 55.8 32.8 75.2L256 430.7 431.2 268c20.9-19.4 32.8-46.7 32.8-75.2v-3.3c0-47.3-33.6-88-80.1-96.9c-34-6.5-69 5.4-92 31.2c0 0 0 0-.1 .1s0 0-.1 .1l-17.8 20c-.3 .4-.7 .7-1 1.1c-4.5 4.5-10.6 7-16.9 7s-12.4-2.5-16.9-7z" />
         </HeartSvg>
       )}
+      <LikeCountHeart>{item.likeCount}</LikeCountHeart>
     </HeartDiv>
   );
 };
 
+export default Like;
+
 // 좋아요 icon
 const HeartDiv = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
   position: absolute;
-  top: 1em;
-  left: 80%;
+  top: 10px;
+  right: 10px;
   color: #fff;
 `;
 const HeartSvg = styled.svg`
@@ -85,7 +101,4 @@ const HeartSvg = styled.svg`
 
 const LikeCountHeart = styled.span`
   font-size: 10px;
-  left: 90%;
 `;
-
-export default Like;
