--- src/components/PostLike.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const PostLike = () => {
-  return (
-    <div>PostLike</div>
-  )
-}
-
-export default PostLike
